fix broken catalog styles on mobile
fix style product header padding on mobile
fix style filter padding
fix dropdown zindex sort
fix grid for 2 products
fix position of elements
