+++ src/components/Products/Dropdown.module.scss
@@ -11,7 +11,7 @@
   border: 2px solid #eaecf0;
   border-radius: 50px;
   position: relative;
-  z-index: 999;
+  z-index: 1;
   > p {
     line-height: 24px;
     width: 100%;
