+++ src/components/Products/Header.tsx
@@ -1,7 +1,9 @@
-import { useState } from 'react';
+import { useContext, useState } from 'react';
 import { SortDropdown } from './SortDropdown';
 import styles from './Products.module.scss';
 import FilterTags from './FilterTags';
+import { mobileFilter } from '../../assets/catalog';
+import { ProductFiltersContext } from '../../pages/Catalog/CategoryProducts';
 
 const Header = ({ title }: { title: string }) => {
   const [sortByFilter, setSortByFilter] = useState<string[]>(['Low To High']);
@@ -23,8 +25,32 @@ const Header = ({ title }: { title: string }) => {
         </div>
       </div>
 
+      <div className={styles.mobile_devider}></div>
+
+      <MoreFiltersMobileBtn />
+
       <FilterTags />
     </>
+  );
+};
+
+const MoreFiltersMobileBtn = () => {
+  const { setIsMobileFilterVisible } = useContext(ProductFiltersContext)!;
+
+  const handleVisibleFilter = () => {
+    setIsMobileFilterVisible(true);
+  };
+
+  return (
+    <div className={styles.mobile_filter}>
+      <span>Filter:</span>
+
+      <img
+        src={mobileFilter}
+        alt='mobile filter icon'
+        onClick={handleVisibleFilter}
+      />
+    </div>
   );
 };
